sw: increase horizontal line parastyle default borderwidth
RES_POOLCOLL_HTML_HR had the absolute minimum width of 1 (0.05pt).
Dividing by 3 in the "convert to word size" formula returned zero.
It looks like import or export must increase the size to 3 because
round-tripping in ODT format loads in a larger value that works
when resaving as .doc.

cui/source/tabpages/border.cxx sets the default for a double
style border to 10, but the minimal value of 3 will be better
since it matches the ODT round-tripped value (and looks more
like a single line which people probably expect from the
horizontal line).

Change-Id: I32a90f2add3fe746359cd9cece14762ccc2a58c1
Reviewed-on: https://gerrit.libreoffice.org/61022
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>