tdf#111980 oox: connect empty radio groups
affects doc, docx, xls, xlsx (at least). Handling optionbutton
groups is new since 2017, so incremental enhancements are OK.

In LO, if an option button with an empty groupName (like any form
control which has no groupName option), then it stands alone,
which is a ridiculous position to be in for a radio button.
So, lets put them all together into an automatic group.

In MSO, buttons without any explicit group stick together,
so the effective group is the entire document, or entire sheet.

One complication is that ActiveX radio buttons don't
interact with Form radio buttons, even if all of them
have no group name. That situation is not handled by
this patch (and not very likely to happen since ActiveX
defaults the group name to the sheet name).

(A Group Box can also be used, which automatically groups
all buttons fully within its area. That will be much harder
to implement since every object will need to first be
categorized as groupbox or optionControl - meaning that everything
must first be imported. That implies a function call from every
application doing the import.)

Change-Id: I204adaea75ce99999a16c0cc4e9c3b8b08da9433
Reviewed-on: https://gerrit.libreoffice.org/61064
Tested-by: Jenkins
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>