tdf#111980 oox optionbutton autoGroup inside GroupBox
The area of a GroupBox indicates which radio buttons
are considered to be part of the same group. The
button needs to be fully inside of the groupbox
in order to participate. This patch resolves the last
worry of commit 21f9d144a5e30fe969f9558f3ca6d5d85afbd915

Change-Id: Ie6057337c63bf9eb173a0615e30c8d4e4d0c7a19
Reviewed-on: https://gerrit.libreoffice.org/61131
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>