tdf#120174 : Always copy mbValueCached flag in Assign()
so that fast result return maintains correctness.

Although not necessary for fixing this bug,
lets take advantage of fast result return in the case
of empty result token by setting mfValue to 0.0 and
setting mbValueCached flag.

The fast result return was implemented in :-

<commit-header>
commit f7771de0e067a0471806c04ec9ac884394cc0331
Author: Dennis Francis <dennis.francis@collabora.co.uk>
Date:   Wed Jul 11 17:19:28 2018 +0530

    Allow fast result return for formula-cells with...

    "double" result which is a very frequent use-case.
    This improves overall running times in most cases,
    not just for the threaded path.

    Change-Id: I18d10ee3cea613923e2057e746a6a8187bb18647
    Reviewed-on: https://gerrit.libreoffice.org/59395
    Tested-by: Jenkins
    Reviewed-by: Luboš Luňák <l.lunak@collabora.com>
</commit-header>

Also added a ui-test for the fix. A normal unit test would
have sufficed, but I could not get it to reproduce the issue
from a ucalc environment somehow.

Change-Id: I9b95b571b596a7bb68768ea0de8ee4334448b540
Reviewed-on: https://gerrit.libreoffice.org/61140
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Jenkins