sw: rdf: Split graph-name lookup from getStatement
The graph-name lookup is significantly costly (compared
to the statement lookup, esp. when no statements exist).
Luckily, the graph-names do not change often and in the
course of enumerating all paragraphs (as happens for
paragraph-signature validation) it doesn't change at all.

This split allows for doing the graph-name lookup only
once and also allows for passing custom graph-names
directly, if we know them already.

Change-Id: I75425df201becb41105ba1fa6ba580af202d035c
Reviewed-on: https://gerrit.libreoffice.org/63002
Tested-by: Jenkins
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>