tdf#119829 use font cache based glyph rect cache
The current glyph cache on Windows didn't work for multiple fonts
and had to be manually invalidated on font change. The new glyph
rect cache is coupled to the font cache and will invalidate all
cached glyph rects, if a cached font is released.

So switch to the new cache on Windows.

Change-Id: Ic641f78e2e664dc0ad52e595d5fdfb6ef611eb3f
Reviewed-on: https://gerrit.libreoffice.org/61278
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>