bodge up a workaround for GtkComboBox menus not appearing under wayland
when the menu is too tall to fit and the active entry is sufficiently
far down the menu that the problem gets triggered.

i.e. https://gitlab.gnome.org/GNOME/gtk/issues/94

as a workaround set wrap to 1, which is good enough to make
GtkComboBox take an alternative path to the one which sets
the problematic rect-anchor-dy value and the menu will at least
appear

Change-Id: Ia05887e0db727b7e1045179e36d348d22228a0ce
Reviewed-on: https://gerrit.libreoffice.org/61416
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>