sw_redlinehide_3: SwAccessibleHyperlink listen at SwFormatINetFormat
There is an annoying problem here: how does a SwAccessibleHyperlink
check whether its corresponding model hint still exists or not?

The existing check via the hint-position appears not ideal, and with the
MergedPara the obvious extension is to add a SwTextNode pointer (or
node index?) which would make it even less ideal...

Instead, do as a lot of classes in unocore do, and register as a
listener on the pool item.

There is another aspect there in how the SwAccessibleHyperTextData
map gets disposed quite eagerly when anything changes, but the
SwAccessibleHyperlink can live longer than this map.

Change-Id: I7d0780cf28794b4ef68cff5d640190694f67530e