Make LibreOfficeLight build again
(And it actually works now again, as far as I see, after the recent
fix to LibreOfficeKit's iOS code.)

Adapt to earlier changes: The generated files are now in workdir, not
in the ios source directory.

Use org.libreoffice.ios.LibreOfficeLight for the
PRODUCT_BUNDLE_IDENTIFIER instead of janI's own.

(Additionally the DEVELOPMENT_TEAM was reset to the one I use;
apparently there is no way to make sure that developer-specific
setting is in a file not in version control?)

Change-Id: I575561583f584b5ac3c759d115b1c9c6dc97ef94