Skip the backup dance on iOS for now
The code is much too convoluted for my little brain. We can't create a
backup copy of a file from outside the sandbox in the same folder as
the original file. At least not using just normal Unix APIs. And if we
store it somewhere else, how would the user find it anyway? Let's just
skipt this mess for now.

No idea how the code manages to create backup files in the same folder
as the actual document in a sandboxed LibreOffice on macOS. Or does
it? Maybe we do some similar bypassing of the backup dance at some
other place in the code already, and I should just have made that
happen for iOS, too?

Change-Id: I0c90edf9e72f54cce78b2cd325e67c710b6df745