sw_redlinehide_3: adapt SwContentTree/SwOutlineContent to gaps
The SwOutlineContent::GetPos() returns the outline position in the
document model / IDocumentOutlineNodesAccess; rename it to make that
more clear.

Also it's already possible to have gaps in the list entries by setting
the outline level so that inner levels are filtered; the
SwContentTree::ExecCommand() doesn't appear to handle this properly,
it needs some tweaks to skip over the outline nodes that don't have list
entries.

The GetYPos() does not contain the outline index, it contains the list
entry index.

Change-Id: Ifae0b9648ce61bb9aa470172ef989b2532370138