tdf#120657 Pie chart saved as DOC doesn't preview in MSO
regression from

    commit 737d640c5367364cf81d5e5884986b7713e7fbff
    add SvStream::TellEnd

the problem is that in this stack trace

    sot/source/sdstor/stgelem.hxx:130
    sot/source/sdstor/stgdir.cxx:300
    sot/source/sdstor/stg.cxx:245
    sot/source/sdstor/storage.cxx:187
    tools/source/stream/stream.cxx:1908
    sw/source/filter/ww8/ww8scan.cxx:6267

The StgEntry returns a size of 0. If we do something else to the
SotStorageStream first, then it materialises a temporary stream, and
that stream returns a valid size.

Also implement suggestions from mike kaganski:

IMO, it would be more robust to implement this new function as virtual
(instead of the remainingSize()), and reimplement the latter as
non-virtual one using the new one and doing the necessary correction.

Change-Id: Id1462d918570a684ec67099c30181436a27d3c6a
Reviewed-on: https://gerrit.libreoffice.org/62984
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>