Simplify and fix Java UNO API test makefiles
Originally I just wanted to add the juh.jar to the list of jars
of the UNO API tests, but this became tedious work, so after the
first few files I decided to replace the similiar makefiles with
a common define for the *_unoapi* tests.

This patch adds two new make defines to be used used by Java UNO
and UNO API tests:
- gb_JunitTest_set_unoapi_test_defaults
- gb_JunitTest_set_unoapi_test_class_and_jars

The first one will deduce most defaults from the test name, but
still allows to optionally override most settings.

If a test doesn't match the default at all, the 2nd define still
shares the common jar files and the main Java UNO class, so the
second define adds these to your makefile.

The real fix is to add juh.jar to gb_JunitTest_use_jars.

Change-Id: I4342fdac5e31f85ea18fb4268e13c287a7adc2b7
Reviewed-on: https://gerrit.libreoffice.org/63118
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>