avoid valgrind uninitialized jump warning with clang
I'm not sure what exactly the actual problem is, but valgrind reports
unitialized reads of mbMulti, although that one is fine and the culprint
is in fact mbCenter. Which happens to be bit 16 of the bitfield, and
the generated assembly looks fishy, just for that one specific bit.
Adding extra padding, extra statements or even just reordering avoids it.
So this is lame, but just reorder it and be done with it, this should
be otherwise harmless and I've spent enough time on this.

Change-Id: I25be55ebe0e9b6dc981793aa7caeae2a40bb27a5
Reviewed-on: https://gerrit.libreoffice.org/63186
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>