Don't bother setting LC_ALL, LC_CTYPE, and LANG on iOS
An iOS app might itself, for testing and debugging purposes, look for
LANG in the environment (potentially passed to it by Xcode; the OS
does not set such an environment variable). It is confusing if that
then gets set during the execution of core code back to a device's
default.

Change-Id: I9dcf44090aed84b55fd4240bda2562026cd8dacb