Don't create temp file in save destination folder when sandboxed
In a sandboxed environment we don't want to attempt to create
temporary files in the same directory where the user has selected an
output file to be stored. The sandboxed process has permission only to
create the specifically named output file that the user chose in the
system file dialog in that directory.

Change-Id: I65274b4c433bd66b436429bea4b287a0505b72c9