tdf#121561: sw: DOCX: add std/stdPr/stdContent around TOC
During export into DOCX from ODT we need to do it
because in this case the TOC title will be recognized
inside MS Word as part of the TOC.

Later we could add support of these keywords in LO import
in order to detect TOC title from DOCX input.

Added unit test for export.

Change-Id: I7135e91dc04d4c0501e6074a046fc473e041f014
Reviewed-on: https://gerrit.libreoffice.org/63786
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>