Make use of SwView::InsertGraphicDlg() in the mobile app case work
Despite the name of the function, when inserting an image that we
already know the file name for, we don't actually need to invoke any
interactive file dialog (file picker). We don't build any file picker
for non-desktop platforms anyway, so make sure this code does not
crash in trying to access a such.

Change-Id: I91cc8c4d9dac26e66f4f852c3d84f5c778c90644