fix svDoubleRef handling in OpenCL
Basically just adjust it based on svSingleRef handling, as that one
seems sane, while the svDoubleRef case, as a comment puts it, is
"such crack". It didn't seem to make much sense written that way,
and it didn't handle correctly e.g. COUNT() with only strings,
OpCount says it doesn't take strings, but the code passed the arguments
to it anyway, in the "other case".
It'd be better to merge the code into one shared function, but sadly
there are small differencies, so at least this way for now.

Change-Id: Ia5f6ce60dae54b1d5a97e049600503aaa9be3bf0
Reviewed-on: https://gerrit.libreoffice.org/65478
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>