sw_redlinehide_4b: consider the comments in the margin
SwPostItMgr needs to hide the comments that are in delete redlines.

Also notify SwPostItMgr when redlines are created/removed.

SwPostItMgr is owned by the ViewShell but there is currently only one
layout so SwRootFrame still needs to broadcast to all of them.

Also an EndListening call was missing in
SwPostItMgr::CheckForRemovedPostIts() which caused asserts.

Change-Id: Ic0fccde50f6fdaa449afb535476c00c41ba94287