tdf#121987: Don't fail installation if failed to enable WU service
Since commit 79055ee19c175d56d8f7003716595c455b15e163, an attempt to
install UCRT will be performed regardless there is an evidence that it's
present on the system, to workaround some cases where the existing UCRT
is broken (tdf#115405, tdf#119910).

But that made other errors to emerge: on systems where users disable WU
service using some exotic ways, installer is unable to enable the
service, and fails. [1][2] Examples of such hard-disables are using
`sc delete` [3] and associating WU service with a guest account. Many
such cases are reported for Windows 10, where installation of the UCRT
is not required.

So the solution (imperfect, but possibly the best possible here) is to
allow installer to continue in case of failure enabling the service.
This will automatically eliminate all problems related to Win10; and
also for cases where users are advanced enough (the majority of such
hard-disable cases should be those), it might be enough to add a
relevant FAQ entry.

[1] https://ask.libreoffice.org/en/question/172227/cannot-install-631/
[2] https://ask.libreoffice.org/en/question/175571/installation-failed-unknown-error-win10x64/
[3] https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/sc-delete

Change-Id: Ie85016eb6f0667f39412a3089fe1b1855cb1fc73
Reviewed-on: https://gerrit.libreoffice.org/64825
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>