tdf#120342 OSX always lock SolarMutex in drawRect
Since we're now a good OSX citizen and do all our drawing in the
main thread, I believe the workaround from i#93512 and merged in
commit 81ec69125209 ("CWS-TOOLING: integrate CWS i93512_DEV300")
isn't needed anymore. Therefore we can just claim the SolarMutex
and draw. And I couldn't reproduce the deadlock of i#93512 with
this patch applied.

But I already was wrong a few times and many drawing semantics
have changed for OSX 10.14, so I might be wrong again ;-)

Change-Id: Ibbf1c1f394038ee5051bc16d2f3c677f4231b2ba
Reviewed-on: https://gerrit.libreoffice.org/65009
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>